refactor: 全面重构——禁用STL容器，GUI改用Qt5
数据层（满足题目要求：不使用STL容器类）：
- 新增 DynArray<T>  替代 std::vector
- 新增 HashMap<K,V> 替代 std::map/unordered_map
- 新增 MinHeap<T,Cmp> 替代 std::priority_queue
- 新增 Queue<T>     替代 std::queue
- Graph/Dijkstra/TopoSort/OrderManager/FileManager
  全部改写，仅保留 std::string 和 I/O 流

GUI层（Qt5 Widgets）：
- 删除 SFML 依赖
- 新增 MainWindow：左侧深色面板+按钮导航+操作日志
- 新增 GraphWidget：QPainter 自绘路网图，带路径高亮
- 所有操作改用 QDialog 表单，体验更好
- 支持 make qt-gui 一键构建

+++ main.cpp
@@ -61,7 +61,7 @@ static void printPathResult(const PathResult& r, const string& label = "") {
         return;
     }
     cout << "  路径：";
-    for (size_t i = 0; i < r.path.size(); ++i) {
+    for (int i = 0; i < r.path.size(); ++i) {
         int id = r.path[i];
         const Node* n = g.findNode(id);
         cout << id << "(" << (n ? n->name : "?") << ")";
@@ -251,7 +251,7 @@ static void menuDelivery() {
                 cout << "\n";
             } else {
                 cout << "\n批次配送顺序（拓扑排序结果）：\n  ";
-                for (size_t i = 0; i < topoRes.order.size(); ++i) {
+                for (int i = 0; i < topoRes.order.size(); ++i) {
                     int id = topoRes.order[i];
                     const Node* n = g.findNode(id);
                     cout << id << "(" << (n ? n->name : "?") << ")";
